fix: sync before every action
this reduces gas costs. If the sync had happened in the middle of every action after the call to totalAssets happen, that requires all the calculations to get the smoothedout profit and then the actual sync to write that profit does the calculation over again. Now that the sync happens before every action, the lastSyncedTime when totalAssets is called is the block.timestamp and it returns early and saves gas